pyproject.toml
[project]
name = "workspace"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "datasets>=3.5.1",
    "einops>=0.8.1",
    "hydra-core>=1.3.2",
    "numpy>=2.2.5",
    "packaging>=25.0",
    "tensorboard>=2.19.0",
    "timm>=1.0.15",
    "torch>=2.7.0",
    "torchaudio>=2.7.0",
    "torchvision>=0.22.0",
    "tqdm>=4.67.1",
    "transformers>=4.51.3",
    "xlrd>=2.0.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
